% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{get_final_counts}
\alias{get_final_counts}
\title{Regresa los conteos finales para las gr치ficas por Dia/Hora.}
\usage{
get_final_counts(datos, user_input, tipo)
}
\arguments{
\item{datos}{Un tibble con los datos.}

\item{user_input}{el objeto con los par치metros del usuario.}

\item{tipo}{"PGJ"; "SSC"; "C5"; "AXA" Cu치l base de datos.}
}
\value{
tibble con columnas con los eventos agregados por dia.
}
\description{
Regresa los conteos finales para las gr치ficas por Dia/Hora.
}
\examples{
count_final <- get_final_counts(bd$tmp_pgj, input, "PGJ")
}
