% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_server.R
\name{\%|NA|\%}
\alias{\%|NA|\%}
\title{If x is `NA`, return y, otherwise return x}
\usage{
x \%|NA|\% y
}
\arguments{
\item{x, y}{Two elements to test, one potentially `NA`}
}
\description{
If x is `NA`, return y, otherwise return x
}
\examples{
NA \%||\% 1
}
