% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_graficas.R
\name{horas_graf}
\alias{horas_graf}
\title{Grafica Horas Function}
\usage{
horas_graf(dataframe_rec_in, input)
}
\arguments{
\item{dataframe_rec_in}{The reactive function that returns a dataframe with 
the "incidentes viales"}

\item{input}{shiny parameter where the inputs from the UI are store 
and the month or day graph is selected}
}
\value{
The render plot selected in the UI
}
\description{
Generates the graph of the "Incidentes viales" by the selected interval day
}
