% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_bar.R
\name{mod_bar_server}
\alias{mod_bar_server}
\title{Bar Server Function}
\usage{
mod_bar_server(input, output, session)
}
\arguments{
\item{input}{shiny internal}

\item{output}{shiny internal}

\item{session}{shiny internal}
}
\value{
Reactive function thatreturns a vector with a time interval (min, max) dates
}
\description{
The function returns a reactive function that returns a vectos with the date of the interval
}
