% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_csvFileUI.R
\name{mod_csvFileUI_server}
\alias{mod_csvFileUI_server}
\title{csvFileUI Server Function}
\usage{
mod_csvFileUI_server(input, output, session)
}
\arguments{
\item{input}{shiny Parameter where the inputs from the UI are store}

\item{output}{shiny parameter where the id are reference to display in the UI}

\item{session}{shiny parameter to store the session}
}
\description{
Function that loads the new files for to create a new Dataframe to be use by 
the "visualizador de accidentes"
}
