% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_csvFileUI.R
\name{mod_csvFileUI_ui}
\alias{mod_csvFileUI_ui}
\title{EcsvFileUI UI Function}
\usage{
mod_csvFileUI_ui(id, label = "Selecciona el archivo")
}
\arguments{
\item{id}{}
}
\description{
Function that generates the user interface to load the files 
to create the database that is use to visualize "accidentes viales"
}
