% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_DBSelector.R
\name{mod_DBSelector_ui}
\alias{mod_DBSelector_ui}
\alias{mod_DBSelector_server}
\title{DBSelector UI Function}
\usage{
mod_DBSelector_ui(id)

mod_DBSelector_server(input, output, session, interval_ba_rea)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
