% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_DBSelector.R
\name{mod_DBSelector_server}
\alias{mod_DBSelector_server}
\title{DBSelector Server Function}
\usage{
mod_DBSelector_server(input, output, session, interval_ba_rea)
}
\arguments{
\item{input}{shiny internal}

\item{output}{shiny  internal}

\item{session}{shiny internal}

\item{interval_ba_rea}{Interval reactive that returne time interval to select data}
}
\value{
reactive function that returns a filtered dataframe using the inputs from the UI
}
\description{
Using the filters selected in the UI the function returns a reactive function that returns 
a dataframe filtered using the corresponding filters.
}
\keyword{internal}
