% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_graficas.R
\name{mod_graficas_ui}
\alias{mod_graficas_ui}
\title{graficas UI Function}
\usage{
mod_graficas_ui(id)
}
\arguments{
\item{id}{shiny parameter for the different sessions}
}
\description{
UI function that produces the graphic interface to select the type of plots
and the interval to generate them.
}
