% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_infoBdUI.R
\name{mod_infoBdUI_server}
\alias{mod_infoBdUI_server}
\title{infoBdUI Server Function}
\usage{
mod_infoBdUI_server(input, output, session, parent)
}
\arguments{
\item{input}{shiny Parameter where the inputs from the UI are store}

\item{output}{shiny parameter where the id are reference to display in the UI}

\item{session}{shiny parameter to store the session}
}
\description{
The function that controls the tabs
}
