% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_infoBdUI.R
\name{mod_infoBdUI_ui}
\alias{mod_infoBdUI_ui}
\title{infoBdUI UI Function}
\usage{
mod_infoBdUI_ui(id)
}
\description{
Function that generates the user interface (UI) where the 
information about data sources is display
}
