% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mapa.R
\name{mod_mapa_server}
\alias{mod_mapa_server}
\title{mapa Server Function}
\usage{
mod_mapa_server(input, output, session)
}
\arguments{
\item{input}{shiny Parameter where the inputs from the UI are store}

\item{output}{shiny parameter where the id are reference to display in the UI}

\item{session}{shiny parameter to store the session}
}
\description{
Map server gets the map from openstreetmaps an zoom it acordandly
}
