% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{preprocesa_pgj}
\alias{preprocesa_pgj}
\title{Preprocesa archivo csv de Fiscalía}
\usage{
preprocesa_pgj(pgj)
}
\arguments{
\item{Tibble}{con los datos de fiscalía leídos de csv}
}
\value{
Tibble Con los datos preprocesados listos para utilizarse en la plataforma.
}
\description{
Preprocesa archivo csv de Fiscalía
}
\examples{
pgj_procesado <- preprocesa_pgj(pgj)
}
