% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{preprocesa_pgj_origin}
\alias{preprocesa_pgj_origin}
\title{Preprocesa archivo csv de Fiscalía}
\usage{
preprocesa_pgj_origin(fgj_new)
}
\arguments{
\item{fgj_new}{tabla con los datos de fiscalía leídos de csv}
}
\value{
Tabla con los datos preprocesados listos para utilizarse 
en la plataforma.
}
\description{
Lee el rds original y permite incorporar los nuevos datos usando la fecha
}
