% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_DBSelector.R
\name{mod_DBSelector_ui}
\alias{mod_DBSelector_ui}
\title{DBSelector UI Function}
\usage{
mod_DBSelector_ui(id)
}
\arguments{
\item{id}{shiny id}
}
\description{
Generate the UI for selecting the "Incidentes viales" data source,
type of accident and location
}
