% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{preprocesa_axa}
\alias{preprocesa_axa}
\title{Preprocesa archivo csv de AXA}
\usage{
preprocesa_axa(axa)
}
\arguments{
\item{Tibble}{con los datos de AXA leídos de csv}
}
\value{
Tibble Con los datos preprocesados listos para utilizarse en la plataforma.
}
\description{
Preprocesa archivo csv de AXA
}
\examples{
axa_procesado <- preprocesa_axa(axa)
}
