% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{preprocesa_ssc}
\alias{preprocesa_ssc}
\title{Preprocesa archivo csv de Secretaría de Seguridad Ciudadana}
\usage{
preprocesa_ssc(ssc)
}
\arguments{
\item{ssc}{Tabla con los datos de SSC leídos de csv}
}
\value{
Tabla con los datos preprocesados listos para utilizarse
 en la plataforma.
}
\description{
Preprocesa archivo csv de Secretaría de Seguridad Ciudadana
}
